switch the check for aggregation passes to use the SQL views

+++ pkg/jobrunaggregator/jobrunaggregatorapi/types.go
@@ -20,8 +20,8 @@ type AggregatedTestRunRow struct {
 	PassCount            int
 	FailCount            int
 	FlakeCount           int
-	PassPercentage       int
-	WorkingPercentage    int
+	PassPercentage       float64
+	WorkingPercentage    float64
 	DominantCluster      string
 	//JobLabels            []string
 }
